requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.30.0
pgvector>=0.3.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-multipart>=0.0.9
greenlet>=3.0.0
aiofiles>=24.1.0
